requirements.txt
# LLM Cybersecurity Platform Dependencies

# Core LLM
openai==1.54.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Data Processing
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1

# Machine Learning
scikit-learn==1.5.2
transformers==4.46.3
torch==2.5.1

# Security Tools
python-magic==0.4.27
yara-python==4.5.1
pefile==2024.8.26
capstone==5.0.1  # Disassembly
requests==2.32.3
cryptography==43.0.3

# Network Analysis
scapy==2.6.1
dpkt==1.9.8

# Code Analysis
ast-grep==0.28.0
pylint==3.3.1
bandit==1.8.0  # Security linter
semgrep==1.94.0  # Static analysis

# Blockchain
web3==7.4.0
eth-utils==5.1.0
py-solc-x==2.0.3

# Cloud SDKs
boto3==1.35.63  # AWS
google-cloud-storage==2.18.2  # GCP
azure-storage-blob==12.24.0  # Azure
kubernetes==31.0.0

# Database
sqlalchemy==2.0.36
redis==5.2.0
pymongo==4.10.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.27.2  # For testing FastAPI

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
python-json-logger==3.2.1
colorlog==6.9.0
tqdm==4.67.1
click==8.1.7

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.48

# Code Quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1

# Monitoring & Observability
prometheus-client==0.21.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2

# Rate Limiting & Caching
aiohttp==3.11.9
tenacity==9.0.0  # Retry logic
cachetools==5.5.0

# Email Analysis
email-validator==2.2.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Logging
loguru==0.7.3

# Environment
python-decouple==3.8

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Additional Security
python-nmap==0.7.1
paramiko==3.5.0  # SSH
pyshark==0.6  # Wireshark wrapper

# IoT
pyserial==3.5
bluepy==1.3.0

# Type Hints
typing-extensions==4.12.2

# Async Support
asyncio==3.4.3
aiofiles==24.1.0
